Created docker-compose configuration
Involved minor changes to other Docker files and the superuser script
to fail nicely if the user already exists, so processing can continue.
NB. the docker-compose config builds the whole structure from scratch -
db, backend, frontend. It requires the frontend code to be checked out
in an adjacent directory to the backend code and using the same folder
name as specified in Github.